+++ backend/canisters/community/api/src/updates/mod.rs
@@ -37,6 +37,7 @@ pub mod create_channel;
 pub mod create_user_group;
 pub mod decline_invitation;
 pub mod delete_channel;
+pub mod delete_channel_history;
 pub mod delete_messages;
 pub mod delete_user_groups;
 pub mod delete_webhook;